requirements.txt
streamlit==1.28.1
langchain-openai==0.0.5
python-dotenv==1.0.0
requests==2.31.0
Pillow>=10.0.0

# Azure 서비스 (이미지 분석 및 검색 기능)
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.35.1
isodate==0.7.2

# 선택사항 (해당 기능 사용 시에만 설치)
# azure-ai-search==11.4.0b11
# langchain-community==0.0.10
